put tests in the right spot, add tests for task hooks
